requirements.txt
# backend/requirements.txt
Django>=3.2.0,<3.3.0
djangorestframework>=3.12.0,<3.13.0
psycopg2-binary>=2.8.6,<2.9.0
django-cors-headers>=3.7.0,<3.8.0
Pillow>=8.2.0,<8.3.0
openai>=0.27.0
python-dotenv>=0.19.0
dj-database-url>=0.5.0

